% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_2D.R
\name{dice_coef}
\alias{dice_coef}
\title{2D Dice coefficient (F1 score) for single label.}
\usage{
dice_coef(y_true, y_pred, smooth = 1.0)

bce_dice_loss(y_true, y_pred)
}
\description{
Dice coefficient and loss function for the compile of model.
}
\examples{
\dontrun{
# Compile
model <- model \%>\%
  compile(optimizer = optimizer_rmsprop( lr = 0.01 ),
  loss = bce_dice_loss,
  metrics = custom_metric( "dice_coef", dice_coef ))
}
}
\author{
Satoshi Kume
}
