% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.gif.R
\name{Display.gif}
\alias{Display.gif}
\title{Visualization of GIF animation.}
\usage{
Display.GIF(GifFileName, View=TRUE)
}
\arguments{
\item{GifFileName}{a file path of GIF file.}

\item{View}{logical; if TRUE, display the file now.}
}
\value{
Nothing files
}
\description{
Display the GIF animation file.
}
\examples{
\dontrun{

Img <- Display.GIF(GifFileName="./animation.gif", View=TRUE)

}
}
\author{
Satoshi Kume
}
