% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageView2D.R
\name{ImageView2D}
\alias{ImageView2D}
\title{An image visualization function for the 4D array.}
\usage{
ImageView2D(
  ImgArray_x,
  ImgArray_y,
  ImgN = 1,
  lab = c("Original", "Overlay", "Prediction")
)
}
\arguments{
\item{ImgArray_x}{a 4D array}

\item{ImgArray_y}{a 4D array}

\item{ImgN}{a slice of images}

\item{lab}{a label information; three elements of a vector}
}
\description{
An image visualization function for the 4D array.
}
\author{
Satoshi Kume
}
