% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageView2D_pred.R
\name{ImageView2D_pred}
\alias{ImageView2D_pred}
\title{An image visualization function after the model prediction.}
\usage{
ImageView2D_pred(
  ImgArray_x,
  ImgArray_y,
  ImgArray_pred,
  ImgN = 1,
  lab = c("Ground truth", "Prediction", "Overlay", "Binary"),
  SaveAll = FALSE
)
}
\arguments{
\item{ImgArray_x}{a 4D array}

\item{ImgArray_y}{a 4D array}

\item{ImgArray_pred}{a 4D array}

\item{ImgN}{a slice of images}

\item{lab}{a label information; three elements of a vector}

\item{SaveAll}{a logical; locally save the all results ot not.}
}
\description{
An image visualization function after the model prediction.
}
\author{
Satoshi Kume
}
