% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Img2DClustering.R
\name{Img2DClustering}
\alias{Img2DClustering}
\title{Img2DClustering}
\usage{
Img2DClustering(x, Cluster = 3, XY = 1024, OriginalScale = TRUE)
}
\arguments{
\item{Cluster}{a number of cluster}

\item{XY}{pixels values of width and height}

\item{OriginalScale}{}
}
\description{
Img2DClustering
}
\author{
Satoshi Kume
}
