% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iou_2D.R
\name{iou}
\alias{iou}
\title{2D IoU (Intersection-Over-Union) for single label}
\usage{
iou(y_true, y_pred, smooth = 1.0)

iou_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{XXX}

\item{y_pred}{XXX}

\item{smooth}{XXX}
}
\description{
2D IoU (Intersection-Over-Union) for single label
}
\author{
Satoshi Kume
}
