% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2tensor.R
\name{list4tensor}
\alias{list4tensor}
\title{Data conversion of list to tensor.}
\usage{
list4tensor(yList)
}
\arguments{
\item{yList}{list data of 4D array.}
}
\value{
5D Array.
}
\description{
Data conversion of list to tensor.
}
\examples{
\dontrun{

Res4 <- list4tensor(yList)

}
}
\author{
Satoshi Kume
}
