% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningRateSchedule.R
\name{lr_schedule}
\alias{lr_schedule}
\title{Create learning rate scheduler.}
\usage{
lr_schedule(epoch, lr_rate = 0.01, epoch_N = 10, M = c(10, 100, 100))
}
\arguments{
\item{lr_rate}{learning rate at first.}

\item{epoch_N}{the number of epoch to change learning rate.}

\item{M}{a learning rate factor.}
}
\value{
lr_schedule
}
\description{
this create learning rate scheduler.
}
\examples{
\dontrun{

lr_reducer <- callback_learning_rate_scheduler(lr_schedule(lr_rate=0.01, epoch_N=10))

}
}
\author{
Satoshi Kume
}
