% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pred.R
\name{model.pred}
\alias{model.pred}
\title{Calculate the prediction results with a binary label}
\usage{
model.pred(model, x)
}
\arguments{
\item{model}{a model object in keras}

\item{x}{a 4D array of images}
}
\description{
Calculate the prediction results with a binary label
}
\author{
Satoshi Kume
}
