% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{model_nodes}
\alias{model_nodes}
\title{Network diagram for the keras model object}
\usage{
model_nodes(x)
}
\arguments{
\item{model}{the model object in Keras.}

\item{width}{a plot size of width.}

\item{height}{a plot size of height.}
}
\value{
A plot of the model network.
}
\description{
visualize a model in Keras.
  This function was modified the plot_model function from deepviz (https://github.com/andrie/deepviz).
}
\examples{
\dontrun{
plot_model_modi(model)
}

}
\references{
deepviz (https://github.com/andrie/deepviz)
}
\author{
Satoshi Kume
}
