% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot_modi.R
\name{plot_model_modi}
\alias{plot_model_modi}
\title{Create network view for the keras model object.}
\usage{
plot_model_modi(model, width, height, ...)
}
\arguments{
\item{width}{a plot size of width.}

\item{height}{a plot size of height.}
}
\value{
A plot of the model network.
}
\description{
visualize a model in Keras
}
\examples{
\dontrun{
plot_model_modi(model)
}
}
\author{
Satoshi Kume
}
