% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_2D_image.R
\name{processing_2D_image}
\alias{processing_2D_image}
\title{Create an array data of 3D images}
\usage{
processing_2D_image(
  file,
  shape,
  type = "png",
  filter = "bilinear",
  normalize = FALSE,
  clahe = FALSE,
  GammaVal = 1
)
}
\arguments{
\item{file}{a file path}

\item{shape}{a vector of width and height of the resized image or Null}

\item{type}{an image type (i.e. png, tiff)}

\item{filter}{a filtering method (default method: bilinear)}

\item{normalize}{a logical; Intensity values linear scaling}

\item{clahe}{a logical; Contrast Limited Adaptive Histogram Equalization}

\item{GammaVal}{a numeric of gamma}
}
\description{
Create an array data of 3D images
}
\author{
Satoshi Kume
}
