% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_plot_model.R
\name{Py_plot_model}
\alias{Py_plot_model}
\title{Create a network diagram for the model object.}
\usage{
Py_plot_model(model, file = "Model.png")
}
\arguments{
\item{model}{the model object in Keras.}

\item{file}{a file name to be saved.}
}
\value{
A image file in the current directory.
}
\description{
visualize a model using plot_model in tensorflow
}
\examples{
\dontrun{
py_plot_model(model)
}
}
\author{
Satoshi Kume
}
