% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMiW.R
\name{rasterMiW}
\alias{rasterMiW}
\title{rasterMiW: display as a raster image}
\usage{
rasterMiW(x, c = 75)
}
\arguments{
\item{c}{chroma value in the HCL color description.}
}
\description{
rasterMiW: display as a raster image
}
\author{
Satoshi Kume
}
