% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImageDataset3D.R
\name{readImageDataset3D}
\alias{readImageDataset3D}
\title{Create an array data of 3D images}
\usage{
readImageDataset3D(Directory, SaveAs = NULL)
}
\arguments{
\item{Directory}{a path of dataset directory}

\item{SaveAs}{a character of save file (.Rds).}
}
\description{
Create an array data of 3D images
}
\author{
Satoshi Kume
}
