% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_3D.R
\name{Singlelabel_3d_dice_coef}
\alias{Singlelabel_3d_dice_coef}
\title{Create network view for the keras model object.}
\usage{
Singlelabel_3d_dice_coef(y_true, y_pred, axis = c(1:3), smooth = 1.0)

Singlelabel_3d_bce_dice_loss(y_true, y_pred)

multilabel_dice_coefficientR(y_true, y_pred, smooth = 1.0)
}
\description{
visualize a model in Keras
}
\author{
Satoshi Kume
}
