% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGrayScale.R
\name{toGrayScale}
\alias{toGrayScale}
\title{convert to the gray scaled image}
\usage{
toGrayScale(x, mode)
}
\arguments{
\item{x}{an Image object}

\item{mode}{A character value; luminance or gray.}
}
\description{
convert to the gray scaled image
}
\author{
Satoshi Kume
}
