% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_v01.R
\name{unet2D_v01}
\alias{unet2D_v01}
\title{Cretae a model of 2D U-net (v01)}
\usage{
unet2D_v01(shape, nlevels = 3, nfilters = 16, dropouts = c(0.1, 0.1, 0.3, 0.3))
}
\arguments{
\item{shape}{a 3-element vector: Width, Height, Channels.}
}
\description{
Cretae a model of 2D U-net (v01)
}
\author{
Satoshi Kume
}
