% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crop2D.R
\name{crop2D}
\alias{crop2D}
\title{crop2D: Crop image to array}
\usage{
crop2D(x, XY = 1024, Crop = 2^3)
}
\arguments{
\item{x}{an image obkject of EBImage}

\item{XY}{a pixel of width and height}

\item{Crop}{a numeric: ex. 2^n}
}
\description{
crop2D: Crop image to array
}
\author{
Satoshi Kume
}
