% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Img2DClustering.R
\name{Img2DClustering}
\alias{Img2DClustering}
\title{Img2DClustering}
\usage{
Img2DClustering(x, Cluster = 3, XY = 1024, OriginalScale = TRUE)
}
\arguments{
\item{x}{an Image object of EBImage}

\item{Cluster}{a number of cluster}

\item{XY}{pixels values of width and height}

\item{OriginalScale}{a logical}
}
\description{
This function provides an image segmentation with k-means clustering.
}
\author{
Satoshi Kume
}
