% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iou_2D.R
\name{iou}
\alias{iou}
\title{Create network view for the keras model object.}
\usage{
iou(y_true, y_pred, smooth = 1)
}
\arguments{
\item{model}{the model object in Keras.}

\item{width}{a plot size of width.}

\item{height}{a plot size of height.}
}
\value{
A plot of the model network.
}
\description{
visualize a model in Keras
}
\examples{
\dontrun{
plot_model_modi(model)
}
}
\author{
Satoshi Kume
}
