% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMiW.R
\name{rasterMiW}
\alias{rasterMiW}
\title{rasterMiW: display as a raster image}
\usage{
rasterMiW(res, c = 75, method = "raster")
}
\arguments{
\item{res}{a result of Img2DClustering}

\item{c}{chroma value in the HCL color description.}

\item{method}{select "browser" or "raster"}
}
\description{
rasterMiW: display as a raster image
}
\author{
Satoshi Kume
}
