% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_3D.R
\name{Singlelabel_3d_dice_coef}
\alias{Singlelabel_3d_dice_coef}
\title{Create network view for the keras model object.}
\usage{
Singlelabel_3d_dice_coef(y_true, y_pred, axis = c(1:3), smooth = 1)
}
\arguments{
\item{model}{the model object in Keras.}

\item{width}{a plot size of width.}

\item{height}{a plot size of height.}
}
\value{
A plot of the model network.
}
\description{
visualize a model in Keras
}
\examples{
\dontrun{
plot_model_modi(model)
}
}
\author{
Satoshi Kume
}
